Update symfony methods

+++ src/Controller/ProductController.php
@@ -5,11 +5,9 @@ namespace App\Controller;
 use App\Entity\Product;
 use App\Form\ProductType;
 use App\Repository\ProductRepository;
-use App\Repository\UserRepository;
 use App\Service\PeremptionService;
 use App\Service\PhotoService;
 use App\Service\ReportService;
-use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -19,26 +17,16 @@ use Symfony\Component\Routing\Annotation\Route;
 class ProductController extends AbstractController
 {
 
-    /**
-     * @var UserRepository
-     */
-    private $repository;
-
-    public function __construct(UserRepository $repository, EntityManagerInterface $em)
-    {
-        $this->repository = $repository;
-        $this->em = $em;
-    }
-
     #[Route('/', name: 'app_product_index', methods: ['GET'])]
     public function index(ProductRepository $productRepository, ReportService $reportService, PeremptionService $peremption): Response
     {
+        $user = $this->getUser();
         //on recupère la liste des produits de l'utilisateur
-        $products = $productRepository->findAllProductByDate($this->getUser());
+        $products = $productRepository->findAllProductByDate($user);
         //on recupère les poids et prix totaux que représentent des produits séléctionnés
-        $report = $reportService->getReport($this->getUser(), $products);
+        $report = $reportService->getReport($products);
         // on recupère les dates de peremption en assignant des valeurs en fonction de la date de peremption
-        $peremption->getPeremption($this->getUser());
+        $peremption->getPeremption($user);
 
         return $this->render('product/index.html.twig', [
             'report' => $report,
@@ -78,12 +66,12 @@ class ProductController extends AbstractController
             if ($photo) {
                 $newPhoto->addPhoto($product, $photo);
             }
-            $id = $this->getUser();
-            $product->setUser($id);
+            $user = $this->getUser();
+            $product->setUser($user);
             $productRepository->save($product, true);
             $this->addFlash('success', 'Product successfully created');
 
-            return $this->redirectToRoute('app_product_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_product_index');
         }
 
         return $this->renderForm('product/new.html.twig', [
@@ -93,7 +81,7 @@ class ProductController extends AbstractController
     }
 
     #[Route('/edit/{id<\d+>}', name: 'app_product_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Product $product, ProductRepository $productRepository, PhotoService $modifyPhoto, int $id): Response
+    public function edit(Request $request, Product $product, ProductRepository $productRepository, PhotoService $photoService, int $id): Response
     {
         // en cas d'id n'appartenant pas au user courant, on declare une erreur 404 
         if ($product->getUser() !== $this->getUser()) {
@@ -105,7 +93,7 @@ class ProductController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             //action si click sur BACK
             if ($form->getClickedButton() && 'back' === $form->getClickedButton()->getName()) {
-                return $this->redirectToRoute('app_product_index', [], Response::HTTP_SEE_OTHER);
+                return $this->redirectToRoute('app_product_index');
             }
             //action si click sur SAVE
             if ($form->getClickedButton() && 'save' === $form->getClickedButton()->getName()) {
@@ -113,12 +101,12 @@ class ProductController extends AbstractController
                 $photo = $form['image']->getData();
                 // si il y a une image a ajouter ou a modifier, on la sauve dans BDD et repertoire images
                 if ($photo) {
-                    $modifyPhoto->addPhoto($product, $photo);
+                    $photoService->addPhoto($product, $photo);
                 }
                 $productRepository->save($product, true);
                 $this->addFlash('success', 'Product successfully saved');
 
-                return $this->redirectToRoute('app_product_index', [], Response::HTTP_SEE_OTHER);
+                return $this->redirectToRoute('app_product_index');
             }
             //action si click sur DELETE
             if ($form->getClickedButton() && 'delete' === $form->getClickedButton()->getName()) {
@@ -150,20 +138,23 @@ class ProductController extends AbstractController
     #[Route('/{id<\d+>}', name: 'app_product_delete', methods: ['POST'])]
     public function delete(Request $request, Product $product, ProductRepository $productRepository): Response
     {
-
+        // en cas d'id n'appartenant pas au user courant, on declare une erreur 404 
+        if ($product->getUser() !== $this->getUser()) {
+            throw $this->createNotFoundException('The product does not exist');
+        }
         if ($this->isCsrfTokenValid('delete' . $product->getId(), $request->request->get('_token'))) {
-            $old_image = $product->getPhoto();
-            if ($old_image) {
-                $file_to_delete = $this->getParameter('images_directory') . '/' . $old_image->getName();
-                if ($file_to_delete) {
-                    unlink($file_to_delete);
+            $image = $product->getPhoto();
+            if ($image) {
+                $fileToDelete = $this->getParameter('images_directory') . '/' . $image->getName();
+                if ($fileToDelete) {
+                    unlink($fileToDelete);
                 }
             }
             $productRepository->remove($product, true);
             $this->addFlash('success', 'Product successfully deleted');
         }
 
-        return $this->redirectToRoute('app_product_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_product_index');
     }
 
     #[Route('/search', name: 'app_product_search', methods: ['GET'])]
@@ -174,7 +165,7 @@ class ProductController extends AbstractController
 
         if (strlen($query) < 3) {
             $this->addFlash('error', 'Search must contains at least 3 characters');
-            return $this->redirectToRoute('app_product_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('app_product_index');
         }
         // on récupère les produits qui correspondent à la recherche
         $products = $productRepository->findSearchQuery(
@@ -182,31 +173,14 @@ class ProductController extends AbstractController
             $query
         );
         //on recupère les poids et prix totaux que représentent des produits séléctionnés
-        $report = $reportService->getReport($this->getUser(), $products);
+        $report = $reportService->getReport($products);
         // on recupère les dates de peremption en assignant des valeurs en fonction de la date de peremption
         $peremption->getPeremption($this->getUser());
-
 
         return $this->render('product/searchresult.html.twig', [
             'query' => $query,
             'report' => $report,
             'products' => $products
-        ]);
-    }
-
-    #[Route('/scan', name: 'app_product_scan', methods: ['GET'])]
-    public function scan(ProductRepository $productRepository, ReportService $reportService, PeremptionService $peremption): Response
-    {
-        //on recupère la liste des produits de l'utilisateur
-        $products = $productRepository->findAllProductByDate($this->getUser());
-        //on recupère les poids et prix totaux que représentent des produits séléctionnés
-        $report = $reportService->getReport($this->getUser(), $products);
-        // on recupère les dates de peremption en assignant des valeurs en fonction de la date de peremption
-        $peremption->getPeremption($this->getUser());
-
-        return $this->render('product/scan.html.twig', [
-            'report' => $report,
-            'products' => $products,
         ]);
     }
 }
